Update hàm xóa sản phẩm trong file js/main.js

+++ js/main.js
@@ -61,6 +61,10 @@ function editMobile() {
 } 
 
 function del(index) {
-    store.deleteMobile(index);
-    display(store.getHtml());
+    let confirmDel = confirm("Are you sure?");
+    if(confirmDel) {
+        store.deleteMobile(index);
+        display(store.getHtml());
+    }
+
 }
